타입스크립트 설정 (#1)
