+++ babel.config.json
@@ -11,7 +11,7 @@
         }
       }
     ],
-    ["@babel/preset-react", { "runtime": "automatic", "importSource": "@emotion/react" }]
-  ],
-  "plugins": ["@emotion/babel-plugin"]
+    ["@babel/preset-react", { "runtime": "automatic" }],
+    ["@babel/preset-typescript", { "isTSX": true, "allExtensions": true }]
+  ]
 }
