+++ package.json
@@ -2,28 +2,34 @@
   "name": "react-components",
   "private": true,
   "version": "1.0.0",
-  "main": "src/index.jsx",
+  "main": "src/app/index.tsx",
   "license": "MIT",
   "keywords": [
     "f-lab",
     "react",
     "not cra"
   ],
+  "engines": {
+    "npm": "use pnpm instead of npm",
+    "yarn": "use pnpm instead of yarn",
+    "node": ">=20.0.0"
+  },
+  "packageManager": "pnpm@8.15.5",
   "scripts": {
-    "dev": "NODE_ENV=dev webpack serve --open --config webpack.dev.js",
-    "build": "NODE_ENV=production webpack --config webpack.prod.js",
-    "lint-staged": {
-      "*.{js,jsx}": [
-        "eslint --fix",
-        "prettier --write"
-      ]
-    },
+    "dev": "cross-env NODE_ENV=dev webpack serve --open --config webpack.dev.js",
+    "type-check": "tsc --noEmit",
+    "build": "cross-env NODE_ENV=production webpack --config webpack.prod.js && pnpm run type-check",
+    "preinstall": "npx only-allow pnpm",
     "husky": "husky install",
     "prepare": "husky && husky install"
   },
+  "lint-staged": {
+    "*.{ts,tsx}": [
+      "eslint",
+      "prettier --list-different"
+    ]
+  },
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.0",
     "@vanilla-extract/css": "^1.14.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
@@ -32,10 +38,16 @@
     "@babel/core": "^7.24.1",
     "@babel/preset-env": "^7.24.1",
     "@babel/preset-react": "^7.24.1",
-    "@emotion/babel-plugin": "^11.11.0",
+    "@babel/preset-typescript": "^7.24.1",
+    "@types/jest": "^29.5.12",
+    "@types/react": "^18.2.69",
+    "@types/react-dom": "^18.2.22",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
     "@vanilla-extract/webpack-plugin": "^2.3.7",
     "babel-jest": "^29.7.0",
     "babel-loader": "^9.1.3",
+    "cross-env": "^7.0.3",
     "css-loader": "^6.10.0",
     "dotenv": "^16.4.5",
     "eslint": "^8.57.0",
@@ -52,17 +64,12 @@
     "lint-staged": "^15.2.2",
     "mini-css-extract-plugin": "^2.8.1",
     "prettier": "^3.2.5",
+    "ts-jest": "^29.1.2",
+    "typescript": "^5.4.3",
+    "typescript-eslint": "^7.3.1",
     "webpack": "^5.90.3",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^5.0.4",
-    "webpack-merge": "^5.10.0",
-    "yarn": "^1.22.22"
-  },
-  "engines": {
-    "npm": "use yarn instead of npm",
-    "node": ">=20.0.0"
-  },
-  "lint-staged": {
-    "*.js": "eslint --cache --fix"
+    "webpack-merge": "^5.10.0"
   }
 }
