+++ src/app/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom/client';
 import App from './App';
 
-const container = document.getElementById('root');
+const container = document.getElementById('root')!;
 
 ReactDOM.createRoot(container).render(<App />);
