+++ tsconfig.json
@@ -1,0 +1,23 @@
+{
+  "include": ["src/**/*"],
+  "exclude": [
+    "node_modules",
+    "webpack.*.js",
+    "jest.config.js"
+  ],
+  "compilerOptions": {
+    "target": "ES6",
+    "jsx": "preserve",
+    "module": "ES6",
+    "moduleResolution": "node16",
+    "resolveJsonModule": true,
+    "allowJs": true,
+    "noEmit": true,
+    "noEmitOnError": true,
+    "isolatedModules": true,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true
+  }
+}
