+++ webpack.dev.js
@@ -2,7 +2,16 @@ const { merge } = require('webpack-merge');
 const common = require('./webpack.common');
 
 module.exports = merge(common, {
+  mode: 'development',
   devtool: 'inline-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.css.ts$/,
+        use: ['style-loader', 'css-loader'],
+      },
+    ],
+  },
   devServer: {
     static: './dist',
     port: 3000,
