+++ webpack.prod.js
@@ -1,6 +1,23 @@
 const { merge } = require('webpack-merge');
 const common = require('./webpack.common');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = merge(common, {
-  devtool: 'source-map',
+  mode: 'production',
+  devtool: 'hidden-source-map',
+  module: {
+    rules: [
+      {
+        test: /\.css.ts$/,
+        use: [MiniCssExtractPlugin.loader, 'css-loader'],
+      },
+    ],
+  },
+  plugins: [
+    new MiniCssExtractPlugin({
+      linkType: false,
+      filename: '[name]..css',
+      chunkFilename: '[name].css',
+    }),
+  ],
 });
